+++ packages/client-ledger/RELEASES.md
@@ -127,3 +127,13 @@ Based on:
 - [typescript v1.0.0-alpha.12] packages/client-ledger
 ### Releases
 - [NPM v1.0.0-alpha.12] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.12 - packages/client-ledger
+
+## 2024-08-01 00:17:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.349.2 (2.382.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.13] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.13] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.13 - packages/client-ledger